separated settings from themes so they are imported separately
