+++ architectures/themes/__init__.py
@@ -1,4 +1,6 @@
-class _Settings():
+from architectures.themes.settings import GraphSettings, ClusterSettings, NodeSettings, EdgeSettings
+
+class _Theme():
 
     def __setitem__(self, key, value):
         self.__dict__[key] = value
@@ -11,324 +13,22 @@ class _Settings():
 
     def __iter__(self):
         return iter(self.__dict__.items())
-class GraphSettings(_Settings):
-    
-    def __init__(
-        self,
-        bgcolor = "white",
-        center = "false",
-        charset = "utf-8",
-        clusterrank = "local",
-        colorscheme = "",
-        comment = "",
-        compound = "true",
-        concentrate = "false",
-        fontcolor = "black",
-        fontname = "times-roman",
-        fontpath = "system-dependent",
-        fontsize = "14",
-        forcelabels = "true",
-        gradientangle = "",
-        imagepath = "",
-        label = "",
-        labeljust = "c",
-        labelloc = "b",
-        landscape = "false",
-        layout = "dot",
-        margin = "0",
-        mclimit = "1",
-        newrank = "false",
-        nodesep = "0.25",
-        nojustify = "false",
-        # "nslimit = "",
-        # "nslimit1 = "",
-        ordering = "",
-        orientation = "0",
-        outputorder = "breadthfirst",
-        pack = "false",
-        # "packmode = "node",
-        pad = "0.0555",
-        pagedir = "bl",
-        quantum = "0.0",
-        rankdir = "TB",
-        ranksep = "0.5",
-        ratio = "auto",
-        remincross = "true",
-        rotate = "0",
-        searchsize = "30",
-        showboxes = "0",
-        size = "",
-        sortv = "0",
-        splines = "line",
-        style = "",
-        viewport = "",
-        **kwargs
-    ):
-        self.bgcolor = bgcolor
-        self.center = center
-        self.charset = charset
-        self.clusterrank = clusterrank
-        self.colorscheme = colorscheme
-        self.comment = comment
-        self.compound = compound
-        self.concentrate = concentrate
-        self.fontcolor = fontcolor
-        self.fontname = fontname
-        self.fontpath = fontpath
-        self.fontsize = fontsize
-        self.forcelabels = forcelabels
-        self.gradientangle = gradientangle
-        self.imagepath = imagepath
-        self.label = label
-        self.labeljust = labeljust
-        self.labelloc = labelloc
-        self.landscape = landscape
-        self.layout = layout
-        self.margin = margin
-        self.mclimit = mclimit
-        self.newrank = newrank
-        self.nodesep = nodesep
-        self.nojustify = nojustify
-        # "nslimit = ""
-        # "nslimit1 = ""
-        self.ordering = ordering
-        self.orientation = orientation
-        self.outputorder = outputorder
-        self.pack = pack
-        # "packmode = "node"
-        self.pad = pad
-        self.pagedir = pagedir
-        self.quantum = quantum
-        self.rankdir = rankdir
-        self.ranksep = ranksep
-        self.ratio = ratio
-        self.remincross = remincross
-        self.rotate = rotate
-        self.searchsize = searchsize
-        self.showboxes = showboxes
-        self.size = size
-        self.sortv = sortv
-        self.splines = splines
-        self.style = style
-        self.viewport = viewport
 
-        self.__dict__.update(kwargs)
+    def get_delta_dict(self, first_dict, second_dict):
+        # create a merged dict
+        merged_dict = {**first_dict, **second_dict}
+        # create a new dict
+        new_dict = {}
+        for k, v in merged_dict.items():
+            if k in first_dict:
+                if first_dict[k] != v:
+                    new_dict[k] = v
+            else:
+                new_dict[k] = v
 
+        return new_dict
 
-class ClusterSettings(_Settings):
-    def __init__(
-        self,
-        bgcolor = "transparent",
-        color = "black",
-        colorscheme = "",
-        fillcolor = "black",
-        fontcolor = "black",
-        fontname = "times-roman",
-        fontsize = "14",
-        gradientangle = "",
-        label = "",
-        labeljust = "c",
-        labelloc = "t",
-        layer = "",
-        margin = "8",
-        nojustify = "false",
-        pencolor = "black",
-        penwidth = "1",
-        peripheries = "1",
-        sortv = "0",
-        style = "",
-        **kwargs
-    ):
-        self.bgcolor = bgcolor
-        self.color = color
-        self.colorscheme = colorscheme
-        self.fillcolor = fillcolor
-        self.fontcolor = fontcolor
-        self.fontname = fontname
-        self.fontsize = fontsize
-        self.gradientangle = gradientangle
-        self.label = label
-        self.labeljust = labeljust
-        self.labelloc = labelloc
-        self.layer = layer
-        self.margin = margin
-        self.nojustify = nojustify
-        self.pencolor = pencolor
-        self.penwidth = penwidth
-        self.peripheries = peripheries
-        self.sortv = sortv
-        self.style = style
-
-        self.__dict__.update(kwargs)
-
-
-class NodeSettings(_Settings):
-    def __init__(
-        self,
-        color = "black",
-        colorscheme = "",
-        comment = "",
-        distortion = "0",
-        fillcolor = "lightgrey",
-        fixedsize = "false",
-        fontcolor = "black",
-        fontname = "times-roman",
-        fontsize = "14",
-        gradientangle = "",
-        group = "",
-        height = "0.5",
-        # image = "",
-        imagepos = "tc",
-        imagescale = "false",
-        # label = "",
-        labelloc = "c",
-        layer = "",
-        margin = "0.11,0.055",
-        nojustify = "false",
-        ordering = "",
-        orientation = "0",
-        penwidth = "1",
-        peripheries = "1",
-        pos = "",
-        regular = "false",
-        shape = "ellipse",
-        shapefile = "",
-        showboxes = "0",
-        sides = "4",
-        skew = "0",
-        sortv = "0",
-        style = "",
-        width = "0.75",
-        # xlabel = "",
-        **kwargs
-    ):
-        self.color = color
-        self.colorscheme = colorscheme
-        self.comment = comment
-        self.distortion = distortion
-        self.fillcolor = fillcolor
-        self.fixedsize = fixedsize
-        self.fontcolor = fontcolor
-        self.fontname = fontname
-        self.fontsize = fontsize
-        self.gradientangle = gradientangle
-        self.group = group
-        self.height = height
-        # self.image = image
-        self.imagepos = imagepos
-        self.imagescale = imagescale
-        # self.label = label
-        self.labelloc = labelloc
-        self.layer = layer
-        self.margin = margin
-        self.nojustify = nojustify
-        self.ordering = ordering
-        self.orientation = orientation
-        self.penwidth = penwidth
-        self.peripheries = peripheries
-        self.pos = pos
-        self.regular = regular
-        self.shape = shape
-        self.shapefile = shapefile
-        self.showboxes = showboxes
-        self.sides = sides
-        self.skew = skew
-        self.sortv = sortv
-        self.style = style
-        self.width = width
-        # self.xlabel = xlabel
-
-        self.__dict__.update(kwargs)
-
-
-class EdgeSettings(_Settings):
-    def __init__(
-        self,
-        arrowhead = "normal",
-        arrowsize = "1",
-        arrowtail = "normal",
-        color = "black",
-        colorscheme = "",
-        comment = "",
-        constraint = "true",
-        decorate = "false",
-        dir = "forward",
-        fillcolor = "black",
-        fontcolor = "black",
-        fontname = "times-roman",
-        fontsize = "14",
-        headclip = "true",
-        headlabel = "",
-        headport = "center",
-        label = "",
-        labelangle = "-25",
-        labeldistance = "1",
-        labelfloat = "false",
-        labelfontcolor = "black",
-        labelfontname = "times-roman",
-        labelfontsize = "14",
-        layer = "",
-        lhead = "",
-        ltail = "",
-        minlen = "1",
-        nojustify = "false",
-        penwidth = "1",
-        pos = "",
-        samehead = "",
-        sametail = "",
-        showboxes = "0",
-        style = "",
-        tailclip = "true",
-        taillabel = "",
-        tailport = "center",
-        weight = "1",
-        xlabel = "",
-        **kwargs
-    ):
-        self.arrowhead = arrowhead
-        self.arrowsize =arrowsize
-        self.arrowtail = arrowtail
-        self.color = color
-        self.colorscheme = colorscheme
-        self.comment = comment
-        self.constraint = constraint
-        self.decorate = decorate
-        self.dir = dir
-        self.fillcolor = fillcolor
-        self.fontcolor = fontcolor
-        self.fontname = fontname
-        self.fontsize = fontsize
-        self.headclip = headclip
-        self.headlabel = headlabel
-        self.headport = headport
-        self.label = label
-        self.labelangle = labelangle
-        self.labeldistance = labeldistance
-        self.labelfloat = labelfloat
-        self.labelfontcolor = labelfontcolor
-        self.labelfontname = labelfontname
-        self.labelfontsize = labelfontsize
-        self.layer = layer
-        self.lhead = lhead
-        self.ltail = ltail
-        self.minlen = minlen
-        self.nojustify = nojustify
-        self.penwidth = penwidth
-        self.pos = pos
-        self.samehead = samehead
-        self.sametail = sametail
-        self.showboxes = showboxes
-        self.style = style
-        self.tailclip = tailclip
-        self.taillabel = taillabel
-        self.tailport = tailport
-        self.weight = weight
-        self.xlabel = xlabel
-
-        self.__dict__.update(kwargs)
-
-
-class Default():
+class Default(_Theme):
 
     def __init__(
         self, 
@@ -360,7 +60,7 @@ class Default():
             self.colors = color_settings
 
 
-class LightMode():
+class LightMode(_Theme):
 
     def __init__(
         self, 
@@ -436,21 +136,7 @@ class LightMode():
         if color_settings:
             self.colors = color_settings
 
-    def get_delta_dict(self, first_dict, second_dict):
-        # create a merged dict
-        merged_dict = {**first_dict, **second_dict}
-        # create a new dict
-        new_dict = {}
-        for k, v in merged_dict.items():
-            if k in first_dict:
-                if first_dict[k] != v:
-                    new_dict[k] = v
-            else:
-                new_dict[k] = v
-
-        return new_dict
-
-class DarkMode():
+class DarkMode(_Theme):
 
     def __init__(
         self,
